requirements.txt
flask
python-dotenv
yfinance
pandas
numpy
matplotlib
mplfinance
scipy
TA-Lib
langchain
langchain-openai
langchain-core
langgraph
openai
ipython
Pillow
requests
akshare
finnhub-python